refactor: improve code separation in plp
